requirements.txt
Django==2.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
Pillow==7.2.0
pytz==2020.1
